+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Improve lookup for paths defined in templates/patches.
+
 ## [0.2.10] - 2021-04-13
 
 ### Fixed